Remove the obsolete jsondb
Fixes #261
