+++ .github/workflows/build.yml
@@ -18,7 +18,7 @@ jobs:
       fail-fast: false
       matrix:
         name: [ossl3, fips, release, dynamic, i686, minimal, pqc, log]
-        db: [jsondb, sqlitedb, nssdb] # FIXME: Add "memorydb" when it will be persistent
+        db: [sqlitedb, nssdb] # FIXME: Add "memorydb" when it will be persistent
     container: fedora:latest
     steps:
       - name: Get Date for DNF cache entry
