+++ Cargo.toml
@@ -82,7 +82,6 @@ sshkdf = []
 tlskdf = []
 
 # Databases
-jsondb = ["memorydb"]
 memorydb = ["aes", "hkdf", "pbkdf2"] # for encryption
 sqlitedb = ["dep:rusqlite", "aes", "hkdf", "pbkdf2"]
 nssdb = ["dep:rusqlite", "aes", "hmac", "pbkdf2"]
